swap behaviors between `coder templates pull` with `coder templates export`
